# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Orla BE API",
    "description": "API do backend Orla — catálogo de óculos multi-tenant",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "tenantSlug": {
        "name": "x-tenant-slug",
        "in": "header",
        "required": true,
        "schema": { "type": "string" },
        "description": "Slug do tenant"
      }
    },
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "code": { "type": "string", "example": "VALIDATION_ERROR" },
          "message": { "type": "string", "example": "Validation failed" },
          "errors": { "type": "object", "additionalProperties": { "type": "array", "items": { "type": "string" } } }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "user"] },
          "tenantId": { "type": "string", "format": "uuid" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "tenantId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Brand": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "tenantId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Image": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "variantId": { "type": "string", "format": "uuid" },
          "url": { "type": "string" },
          "isPrimary": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          },
          "401": { "description": "Credenciais inválidas" },
          "422": { "description": "Erro de validação", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationError" } } } }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "Logout",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "responses": {
          "200": {
            "description": "Logout bem-sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Logged out successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Usuário autenticado",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Listar categorias",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "responses": {
          "200": {
            "description": "Lista de categorias",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Category" } } } } } }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Criar categoria (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Categoria criada" }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": ["Categories"],
        "summary": "Atualizar categoria (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Categoria atualizada" },
          "404": { "description": "Não encontrada" }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Excluir categoria (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Categoria excluída" }
        }
      }
    },
    "/brands": {
      "get": {
        "tags": ["Brands"],
        "summary": "Listar marcas",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "responses": {
          "200": {
            "description": "Lista de marcas",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Brand" } } } } } }
          }
        }
      },
      "post": {
        "tags": ["Brands"],
        "summary": "Criar marca (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/tenantSlug" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Marca criada" }
        }
      }
    },
    "/brands/{id}": {
      "put": {
        "tags": ["Brands"],
        "summary": "Atualizar marca (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "minLength": 1, "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Marca atualizada" }
        }
      },
      "delete": {
        "tags": ["Brands"],
        "summary": "Excluir marca (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Marca excluída" }
        }
      }
    },
...
    "/variants/{variantId}/images": {
      "post": {
        "tags": ["Images"],
        "summary": "Upload de imagem (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "variantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary", "description": "Imagem (max 10MB)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Imagem enviada",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Image" } } } } }
          }
        }
      }
    },
    "/variants/{variantId}/images/{imageId}": {
      "delete": {
        "tags": ["Images"],
        "summary": "Excluir imagem (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "variantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "imageId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Imagem excluída" }
        }
      }
    },
    "/variants/{variantId}/images/{imageId}/primary": {
      "patch": {
        "tags": ["Images"],
        "summary": "Definir imagem como principal (admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/tenantSlug" },
          { "name": "variantId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "imageId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Imagem definida como principal",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Image" } } } } }
          }
        }
      }
    },
...
  }
}